Refactor analysis report architecture
