+++ src/analysis/__init__.py
@@ -1,5 +1,5 @@
 """Analysis utilities and model evaluation."""
-from . import plots, report, tabulation
+from . import plots, report, tabulation, utils
 from .evaluator import ModelEvaluator
 
-__all__ = ["plots", "report", "tabulation", "ModelEvaluator"]
+__all__ = ["plots", "report", "tabulation", "utils", "ModelEvaluator"]
