+++ src/analysis/plots.py
@@ -12,6 +12,7 @@ import numpy as np
 import seaborn as sns
 from sklearn.metrics import auc, roc_curve
 
+from .utils import tweedie_deviance_residuals
 
 def cumulative_gain_plot(model) -> pd.DataFrame:
     """Return cumulative gain DataFrame and plot cumulative gain curve."""
@@ -341,3 +342,318 @@ def plot_variable_distributions(df: pd.DataFrame, variables: Iterable[str], bins
     plt.tight_layout()
     plt.show()
 
+def plot_residual_fit(
+    df: pd.DataFrame,
+    actual_col: str,
+    predicted_col: str,
+    exposure_col: str | None = None,
+    residual_type: str = "raw",  # "raw", "normalized", "tweedie_deviance"
+    p: float | None = None,      # needed if tweedie
+    num_groups: int = 400,
+    figsize: tuple[int, int] = (8, 6),
+    split_name: str | None = None,
+    title: str | None = None,
+) -> plt.Figure:
+    """
+    Plots mean residuals ±1 std vs fitted for any residual type: raw, normalized, 
+    or tweedie deviance.
+
+    Parameters
+    ----------
+    residual_type : str
+        Type of residual to compute: "raw", "normalized", or "tweedie_deviance".
+    p : float, optional
+        Power parameter for Tweedie (needed if residual_type='tweedie_deviance')
+    """
+    df = df.copy()
+
+    if residual_type == "raw":
+        df["residual"] = df[actual_col] - df[predicted_col]
+
+    elif residual_type == "normalized":
+        df["residual"] = (df[actual_col] - df[predicted_col]) / np.sqrt(np.maximum(df[predicted_col], 1e-6))
+
+    elif residual_type == "tweedie_deviance":
+        if p is None:
+            raise ValueError("Must provide `p` for tweedie_deviance residuals.")
+        df["residual"] = tweedie_deviance_residuals(df[actual_col], df[predicted_col], p)
+
+    else:
+        raise ValueError(f"Unknown residual_type '{residual_type}'. Choose from 'raw', 'normalized', 'tweedie_deviance'.")
+
+    # Bin predictions to stabilize plot
+    df["group"] = pd.qcut(df[predicted_col], num_groups, duplicates="drop")
+
+    # Aggregate
+    grouped = df.groupby("group").agg(
+        count=(predicted_col, "count"),
+        avg_predicted=(predicted_col, "mean"),
+        avg_residual=("residual", "mean"),
+        std_residual=("residual", "std"),
+    ).reset_index()
+
+    group_size = int(np.nanpercentile(grouped["count"], 50))
+
+    # Plot
+    fig, ax = plt.subplots(figsize=figsize)
+    ax.plot(grouped["avg_predicted"], grouped["avg_residual"],
+            marker='o', linestyle='-', color='blue', label="Mean Residual")
+
+    ax.fill_between(
+        grouped["avg_predicted"],
+        grouped["avg_residual"] - grouped["std_residual"],
+        grouped["avg_residual"] + grouped["std_residual"],
+        color='blue', alpha=0.2, label="±1 SD"
+    )
+
+    effective_title = title or f"{residual_type.capitalize()} Residual Fit & Heteroskedasticity Check"
+    effective_title += f"\n(Group Size: {group_size})"
+    if split_name:
+        effective_title += f" ({split_name})"
+
+    ax.set_title(effective_title, fontsize=14)
+    ax.set_xlabel("Average Fitted Value")
+    ax.set_ylabel("Residual")
+    ax.axhline(0, linestyle="--", color="gray")
+    ax.grid(True, linestyle="--", alpha=0.7)
+    ax.legend(frameon=True)
+    plt.tight_layout()
+    return fig
+
+
+# --- Gain and lift charts ---------------------------------------------------
+
+def gain_curve_with_gini(
+    df: pd.DataFrame,
+    actual_col: str,
+    predicted_col: str,
+    exposure_col: str | None = None,
+    figsize: tuple[int, int] = (8, 6),
+    split_name: str | None = None,
+    title: str | None = None,
+) -> plt.Figure:
+    """Return a cumulative gain curve with Gini index."""
+    df = df.copy()
+    if exposure_col:
+        df["weighted_actual"] = df[actual_col] * df[exposure_col]
+        df["exposure"] = df[exposure_col]
+    else:
+        df["weighted_actual"] = df[actual_col]
+        df["exposure"] = 1
+
+    df.sort_values(predicted_col, ascending=False, inplace=True)
+    df["cum_exposure"] = df["exposure"].cumsum() / df["exposure"].sum()
+    df["cum_actual"] = df["weighted_actual"].cumsum() / df["weighted_actual"].sum()
+
+    x = np.insert(df["cum_exposure"].values, 0, 0)
+    y = np.insert(df["cum_actual"].values, 0, 0)
+
+    df_perfect = df.sort_values("weighted_actual", ascending=False).copy()
+    df_perfect["cum_exposure"] = df_perfect["exposure"].cumsum() / df_perfect["exposure"].sum()
+    df_perfect["cum_actual"] = df_perfect["weighted_actual"].cumsum() / df_perfect["weighted_actual"].sum()
+    perfect_x = np.insert(df_perfect["cum_exposure"].values, 0, 0)
+    perfect_y = np.insert(df_perfect["cum_actual"].values, 0, 0)
+
+    gini_model = 2 * (np.trapz(y, x) - 0.5)
+    gini_perfect = 2 * (np.trapz(perfect_y, perfect_x) - 0.5)
+    normalized_gini = gini_model / gini_perfect
+
+    effective_title = title or "Gain Curve / Lorenz"
+    if split_name:
+        effective_title += f" ({split_name})"
+    effective_title += f"\nGini={gini_model:.3f}, Norm Gini={normalized_gini:.3f}"
+
+    fig, ax = plt.subplots(figsize=figsize)
+    ax.plot(x, y, label="Model", color="blue")
+    ax.plot(perfect_x, perfect_y, color="green", label="Perfect")
+    ax.plot([0, 1], [0, 1], "--", color="gray", label="Random")
+    ax.fill_between(x, y, np.linspace(0, 1, len(x)), alpha=0.3)
+    ax.set_title(effective_title, fontsize=14)
+    ax.set_xlabel("Cumulative % of Exposure")
+    ax.set_ylabel("Cumulative % of Loss")
+    ax.legend(frameon=True)
+    ax.grid(True, linestyle="--", alpha=0.7)
+    plt.tight_layout()
+    return fig
+
+
+def lift_chart(
+    df: pd.DataFrame,
+    actual_col: str,
+    predicted_col: str,
+    exposure_col: str | None = None,
+    n_bins: int = 10,
+    figsize: tuple[int, int] = (8, 5),
+    annotate_points: bool = False,
+    title: str | None = None,
+    split_name: str | None = None,
+) -> plt.Figure:
+    """Return a lift chart comparing average actual vs. predicted by deciles."""
+    df = df.copy()
+    if exposure_col:
+        df["weighted_actual"] = df[actual_col] * df[exposure_col]
+        df["exposure"] = df[exposure_col]
+    else:
+        df["weighted_actual"] = df[actual_col]
+        df["exposure"] = 1
+
+    df["bin"] = pd.qcut(df[predicted_col], n_bins, duplicates="drop")
+
+    grouped = (
+        df.groupby("bin")
+        .apply(
+            lambda g: pd.Series({
+                "Avg Actual": np.average(g[actual_col], weights=g["exposure"]),
+                "Avg Predicted": np.average(g[predicted_col], weights=g["exposure"]),
+                "Total Exposure": g["exposure"].sum(),
+            })
+        )
+        .reset_index()
+    )
+
+    effective_title = title or "Lift Chart by Prediction Decile"
+    if split_name:
+        effective_title += f" ({split_name})"
+
+    fig, ax = plt.subplots(figsize=figsize)
+    x_values = np.arange(1, len(grouped) + 1)
+
+    ax.plot(x_values, grouped["Avg Actual"], marker="o", linestyle="-", color="navy", label="Actual")
+    ax.plot(x_values, grouped["Avg Predicted"], marker="x", linestyle="--", color="darkorange", label="Predicted")
+
+    if annotate_points:
+        for x, y in zip(x_values, grouped["Avg Actual"]):
+            ax.annotate(f"{y:.2f}", (x, y), textcoords="offset points", xytext=(0, 5), ha="center", fontsize=8)
+
+    ax.set_title(effective_title, fontsize=14)
+    ax.set_xlabel("Decile of Predicted Pure Premium", fontsize=12)
+    ax.set_ylabel("Average Actual / Predicted Pure Premium", fontsize=12)
+    ax.grid(True, which="both", linestyle="--", linewidth=0.5, alpha=0.7)
+    ax.legend(frameon=True)
+    plt.tight_layout()
+    return fig
+
+
+def crunched_residual_plot(
+    df: pd.DataFrame,
+    actual_col: str,
+    predicted_col: str,
+    exposure_col: str | None = None,
+    num_groups: int = 400,
+    figsize: tuple[int, int] = (8, 5),
+    split_name: str | None = None,
+    title: str | None = None,
+) -> plt.Figure:
+    """Return a plot of average residuals vs. fitted values."""
+    df = df.copy()
+    residual = df[actual_col] - df[predicted_col]
+    df["residual"] = residual
+
+    df["group"] = pd.qcut(df[predicted_col], num_groups, duplicates="drop")
+    grouped = (
+        df.groupby("group")
+        .agg(
+            count=(predicted_col, "count"),
+            avg_predicted=(predicted_col, "mean"),
+            avg_residual=("residual", "mean"),
+        )
+        .reset_index()
+    )
+
+    group_size = int(np.nanpercentile(grouped["count"], 50))
+
+    effective_title = title or "Crunched Residuals vs Fitted"
+    effective_title += f"\n(Group Size: {group_size})"
+    if split_name:
+        effective_title += f" ({split_name})"
+
+    fig, ax = plt.subplots(figsize=figsize)
+    ax.scatter(grouped["avg_predicted"], grouped["avg_residual"], s=10, alpha=0.6, color="darkred")
+    ax.axhline(0, linestyle="--", color="gray")
+    ax.set_title(effective_title, fontsize=14)
+    ax.set_xlabel("Average Fitted Value")
+    ax.set_ylabel("Average Residual")
+    ax.grid(True, linestyle="--", alpha=0.7)
+    plt.tight_layout()
+    return fig
+
+
+def partial_gini_plot(
+    df: pd.DataFrame,
+    actual_col: str,
+    predicted_col: str,
+    *,
+    exposure_col: str | None = None,
+    top_percent: int = 20,
+    figsize: tuple[int, int] = (8, 6),
+    split_name: str | None = None,
+    title: str | None = None,
+) -> plt.Figure:
+    """Return a Lorenz curve highlighting the partial Gini at ``top_percent``."""
+    df = df.copy()
+
+    if exposure_col:
+        df["weighted_actual"] = df[actual_col] * df[exposure_col]
+        df["exposure"] = df[exposure_col]
+    else:
+        df["weighted_actual"] = df[actual_col]
+        df["exposure"] = 1
+
+    df.sort_values(predicted_col, ascending=False, inplace=True)
+    df["cum_exposure"] = df["exposure"].cumsum() / df["exposure"].sum()
+    df["cum_actual"] = df["weighted_actual"].cumsum() / df["weighted_actual"].sum()
+
+    pop = np.insert(df["cum_exposure"].values, 0, 0)
+    cum = np.insert(df["cum_actual"].values, 0, 0)
+
+    df_perfect = df.sort_values("weighted_actual", ascending=False).copy()
+    df_perfect["cum_exposure"] = df_perfect["exposure"].cumsum() / df_perfect["exposure"].sum()
+    df_perfect["cum_actual"] = df_perfect["weighted_actual"].cumsum() / df_perfect["weighted_actual"].sum()
+    perf_pop = np.insert(df_perfect["cum_exposure"].values, 0, 0)
+    perf_cum = np.insert(df_perfect["cum_actual"].values, 0, 0)
+
+    def _partial(population: np.ndarray, cumulative: np.ndarray) -> tuple[float, float, float]:
+        cutoff = top_percent / 100
+        idx = np.searchsorted(population, cutoff, side="right")
+        pop_cut = population[:idx]
+        cum_cut = cumulative[:idx]
+        if len(pop_cut) == 0 or pop_cut[-1] < cutoff:
+            pop_cut = np.append(pop_cut, cutoff)
+            cum_cut = np.append(cum_cut, np.interp(cutoff, population, cumulative))
+        area_under_curve = np.trapz(cum_cut, x=pop_cut)
+        equality_area = (cutoff ** 2) / 2
+        gini = (area_under_curve - equality_area) / (1 - equality_area)
+        return gini, equality_area, area_under_curve
+
+    pgini, _, _ = _partial(pop, cum)
+    pgini_perfect, _, _ = _partial(perf_pop, perf_cum)
+    normalized_pgini = pgini / pgini_perfect if pgini_perfect != 0 else 0
+
+    effective_title = title or "Partial Gini"
+    if split_name:
+        effective_title += f" ({split_name})"
+    effective_title += f"\nRaw and Normalized Partial Gini at Top {top_percent}% = {pgini:.3f}, {normalized_pgini:.3f}"
+
+    fig, ax = plt.subplots(figsize=figsize)
+    ax.plot(pop * 100, cum * 100, label="Model", color="blue")
+    ax.plot(perf_pop * 100, perf_cum * 100, color="green", label="Perfect")
+    ax.plot([0, 100], [0, 100], "--", color="gray", label="Random")
+    ax.axvline(top_percent, color="red", linestyle="--", label=f"Top {top_percent}%")
+
+    cutoff_idx = np.searchsorted(pop, top_percent / 100, side="right")
+    ax.fill_between(
+        pop[:cutoff_idx] * 100,
+        cum[:cutoff_idx] * 100,
+        pop[:cutoff_idx] * 100,
+        color="blue",
+        alpha=0.3,
+        label="Partial Gini Area",
+    )
+
+    ax.set_title(effective_title, fontsize=14)
+    ax.set_xlabel("Cumulative % of Exposure")
+    ax.set_ylabel("Cumulative % of Loss")
+    ax.legend(frameon=True)
+    ax.grid(True, linestyle="--", alpha=0.7)
+    plt.tight_layout()
+    return fig
